--- addon/initializers/pollboy.js
@@ -1,8 +1,0 @@
-export function initialize (app) {
-  app.inject('route', 'pollboy', 'service:pollboy')
-}
-
-export default {
-  name: 'pollboy',
-  initialize
-}
