+++ app/initializers/pollboy.js
@@ -1,1 +1,8 @@
-export {default} from 'ember-pollboy/initializers/pollboy'
+export function initialize (container, app) {
+  app.inject('route', 'pollboy', 'service:pollboy')
+}
+
+export default {
+  name: 'pollboy',
+  initialize
+}
